fail cases for parsePath
this was uncovered when I attempted to simplify makeXKeyTestVec by using a map instead of a fold,
and making the paths explicit, which matches the bip32 spec.

everything that depends on paths of form  m/0'/1/2' now fails.

I think the pattern that breaks it is hardened, followed by non-hardened, followed by hardened.

parsePath fails here.