hw: Add "pubkeys" command for displaying pubkey related to address
The "pubkeys" command is the same as the "list" command, except that
it displays the hex-encoded, compressed public key behind an address,
rather than the address itself. Also: the "pubkeys" command does not
display the balance, as I thought the line width was big enough
already, with both a hex-encoded public key and, optionally, a label.