lookupByPubKey / getindex: Lookup the key index of a pubkey
Sometimes, when you've given out a public key to a client, you
want to be able to figure out the index, from which it was derived,
when the client contacts you again, at some later time.
This commit adds the ability to query the database by looking for
the key that matches a given compressed pubkey, and returns this item.