Add cltvEncodeInt/cltvDecodeInt: encode/decode CLTV timestamps
The OP_CHECKLOCKTIMEVERIFY (CLTV) opcode, which has replaced OP_NOP2,
accepts an argument in the uint32 range. This is encoded on the
stack using the same format as regular stack integers, but with the
length limit increased to 5 bytes, in order to get the full uint32
range from the signed stack integer. This commit adds functions for
encoding and decoding integers to/from the stack, for use with CLTV.