Changes required to use "inproc" ZeroMQ transport + example app (#293)
* factor out runSPVServer into: runSPVServerWithContext & runSPVServer

In order to speak to the server using the ZeroMQ in-memory/inter-thread
"inproc" transport, a shared context between sender and receiver is
required. Factoring out 'runSPVServerWithContext' from 'runSPVServer', and
exporting it via Network.Haskoin.Wallet.Server, allows using haskoin-wallet
as an in-process library, for use in eg. Bitcoin-related backend services
or Desktop/GUI Bitcoin wallets.

* Add example app using runSPVServerWithContext & "inproc" transport