Some minor changes (verifySig/signMsg type sign.) + some feature additions (#294)
* hw: Add "pubkeys" command for displaying pubkey related to address

The "pubkeys" command is the same as the "list" command, except that
it displays the hex-encoded, compressed public key behind an address,
rather than the address itself. Also: the "pubkeys" command does not
display the balance, as I thought the line width was big enough
already, with both a hex-encoded public key and, optionally, a label.

* verifySig + signMsg: also accept compressed & uncompressed pubkeys

* lookupByPubKey / getindex: Lookup the key index of a pubkey

Sometimes, when you've given out a public key to a client, you
want to be able to figure out the index, from which it was derived,
when the client contacts you again, at some later time.
This commit adds the ability to query the database by looking for
the key that matches a given compressed pubkey, and returns this item.

* Add "blockinfo" command: show info about a block from the chain

The "blockinfo" command accepts a list of block hashes, and prints
out information about these blocks, if they are present in the
best chain. As a corner case, when no arguments are passed to
"blockinfo", it returns information about the current best block.

* Add cltvEncodeInt/cltvDecodeInt: encode/decode CLTV timestamps

The OP_CHECKLOCKTIMEVERIFY (CLTV) opcode, which has replaced OP_NOP2,
accepts an argument in the uint32 range. This is encoded on the
stack using the same format as regular stack integers, but with the
length limit increased to 5 bytes, in order to get the full uint32
range from the signed stack integer. This commit adds functions for
encoding and decoding integers to/from the stack, for use with CLTV.

* increase dep upper bound: aeson < 1.1; QuickCheck < 2.10