Generate a bip-39 mnemonic using dice (#304)
This patch adds the "dice" utility command to hw. It is a client side
command that will transform a string of 6 sided dice rolls to a bip-39
mnemonic. You have to call the command as so:

hw dice 123456.....

The command will currently only convert 99 dice rolls (255.9 bits of
entropy) to a 24 word mnemonic.