Use new version of secp256k1-haskell
The new library adds compatibility with the version of libsecp256k1 provided by
Fedora. The Fedora package is a fork of that library from the Bitcoin ABC
project. It computes signature nonces slightly differently, so all tests that
depend on signatures being exactly the same for a combination of message and
private key have been removed.